style(server): rename decorators to maintain consistency
