+++ packages/server/src/shared/api/decorators/ParamDishesFilterDecorator.ts
@@ -1,11 +1,11 @@
 import { DishTags, MealType } from '../../../restaurants/dishes/database';
 import { QueryFilters } from './QueryFiltersDecorator';
 
-const ParamDishFilters = () =>
+const ParamDishesFilter = () =>
   QueryFilters([
     { name: 'city', required: false },
     { name: 'mealType', required: false, enum: MealType, enumName: 'MealTypeEnum', isArray: true },
     { name: 'tags', required: false, enum: DishTags, enumName: 'DishTagEnum', isArray: true },
   ]);
 
-export { ParamDishFilters };
+export { ParamDishesFilter };
