+++ packages/server/src/shared/api/decorators/index.ts
@@ -5,7 +5,7 @@ export * from './ApiObjectIdDecorators';
 export * from './ApiOperationDecorators';
 export * from './ExposeId';
 export * from './PaginationDecorator';
-export * from './ParamDishFiltersDecorator';
-export * from './ParamRestaurantFiltersDecorator';
+export * from './ParamDishesFilterDecorator';
+export * from './ParamRestaurantsFilterDecorator';
 export * from './QueryFiltersDecorator';
 export * from './UrlDecorator';
